Refactor of Controller
