+++ src/main/java/ar/com/tacsutn/grupo1/eventapp/web/Controller.java
@@ -1,4 +1,4 @@
-package ar.com.tacsutn.grupo1.eventapp;
+package ar.com.tacsutn.grupo1.eventapp.web;
 
 import org.springframework.web.bind.annotation.RequestMapping;
