+++ src/main/java/ar/com/tacsutn/grupo1/eventapp/web/EventSumController.java
@@ -1,6 +1,5 @@
 package ar.com.tacsutn.grupo1.eventapp.web;
 
-import ar.com.tacsutn.grupo1.eventapp.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
