+++ src/main/java/ar/com/tacsutn/grupo1/eventapp/web/InterestedController.java
@@ -1,6 +1,5 @@
 package ar.com.tacsutn.grupo1.eventapp.web;
 
-import ar.com.tacsutn.grupo1.eventapp.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
