+++ src/test/java/ar/com/tacsutn/grupo1/eventapp/web/ControllerTest.java
@@ -1,6 +1,5 @@
 package ar.com.tacsutn.grupo1.eventapp.web;
 
-import ar.com.tacsutn.grupo1.eventapp.Controller;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
